+++ dio/test/download_test.dart
@@ -94,23 +94,47 @@ void main() {
     );
   });
 
-  test('download write failed', () async {
+  test('delete on error', () async {
     const savePath = 'test/_download_test.md';
     final f = File(savePath)..createSync(recursive: true);
-    final raf = f.openSync(mode: FileMode.write)..lockSync(FileLock.exclusive);
     expect(f.existsSync(), isTrue);
 
     final dio = Dio()..options.baseUrl = serverUrl.toString();
     await expectLater(
-      dio.download('/download', savePath, deleteOnError: false).catchError((e) {
-        throw (e as DioException).error!;
-      }),
-      throwsA(isA<FileSystemException>()),
+      dio
+          .download(
+            '/download',
+            savePath,
+            deleteOnError: true,
+            onReceiveProgress: (count, total) => throw AssertionError(),
+          )
+          .catchError((e) => throw (e as DioException).error!),
+      throwsA(isA<AssertionError>()),
     );
+    expect(f.existsSync(), isFalse);
+  });
 
-    await expectLater(raf.unlock(), completes);
-    await expectLater(raf.close(), completes);
-    await expectLater(f.delete(), completes);
+  test('delete on cancel', () async {
+    const savePath = 'test/_download_test.md';
+    final f = File(savePath)..createSync(recursive: true);
+    expect(f.existsSync(), isTrue);
+
+    final cancelToken = CancelToken();
+    final dio = Dio()..options.baseUrl = serverUrl.toString();
+    await expectLater(
+      dio
+          .download(
+            '/download',
+            savePath,
+            deleteOnError: true,
+            cancelToken: cancelToken,
+            onReceiveProgress: (count, total) => cancelToken.cancel(),
+          )
+          .catchError((e) => throw (e as DioException).type),
+      throwsA(DioExceptionType.cancel),
+    );
+    await Future.delayed(const Duration(milliseconds: 100));
+    expect(f.existsSync(), isFalse);
   });
 
   test('`savePath` types', () async {